Replaced Koin package with di package. Added viewmodel sugaring. Minor code refactoring
