+++ app/src/main/java/com/akshay/newsapp/NewsApp.kt
@@ -1,7 +1,7 @@
 package com.akshay.newsapp
 
 import android.app.Application
-import com.akshay.newsapp.koin.AppInjector
+import com.akshay.newsapp.di.base.AppInjector
 
 class NewsApp : Application() {
