+++ app/src/main/java/com/akshay/newsapp/di/NewsDatabaseModule.kt
@@ -1,4 +1,4 @@
-package com.akshay.newsapp.koin.modules
+package com.akshay.newsapp.di
 
 import com.akshay.newsapp.db.NewsDatabase
 import org.koin.android.ext.koin.androidApplication
@@ -7,15 +7,12 @@ import org.koin.dsl.module
 /**
  * All database related initializations
  */
-
 val newsDatabaseModule = module {
 
-	// NewsDatabase Instance
 	single {
 		NewsDatabase.buildDefault(androidApplication())
 	}
 
-	// NewsArticleDao instance
 	single {
 		get<NewsDatabase>().newsArticlesDao()
 	}
