+++ app/src/main/java/com/akshay/newsapp/di/ViewModelModule.kt
@@ -1,4 +1,4 @@
-package com.akshay.newsapp.koin.modules
+package com.akshay.newsapp.di
 
 import com.akshay.newsapp.ui.viewmodel.NewsArticleViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
@@ -7,10 +7,8 @@ import org.koin.dsl.module
 /**
  * ViewModel initializations
  */
-
 val viewModelModule = module {
 
-	// NewdArticleViewModel instance
 	viewModel {
 		NewsArticleViewModel(get())
 	}
