+++ app/src/main/java/com/akshay/newsapp/ui/activity/MainActivity.kt
@@ -7,16 +7,16 @@ import com.akshay.newsapp.R
 import com.akshay.newsapp.adapter.NewsArticlesAdapter
 import com.akshay.newsapp.model.ViewState
 import com.akshay.newsapp.ui.viewmodel.NewsArticleViewModel
+import com.akshay.newsapp.utils.getViewModel
 import com.akshay.newsapp.utils.observeNotNull
 import com.akshay.newsapp.utils.toast
 import kotlinx.android.synthetic.main.activity_main.*
 import kotlinx.android.synthetic.main.empty_layout.*
 import kotlinx.android.synthetic.main.progress_layout.*
-import org.koin.androidx.viewmodel.ext.android.viewModel
 
 class MainActivity : AppCompatActivity() {
 
-    private val newsArticleViewModel: NewsArticleViewModel by viewModel()
+    private val newsArticleViewModel: NewsArticleViewModel by getViewModel()
 
     /**
      * Starting point of the activity
